+++ src/test/resources/unit/read_constant_buffer_diagnostic.t
@@ -1,0 +1,12 @@
+#!/usr/bin/env perl
+use strict;
+use warnings;
+use Test::More;
+
+our $buffer;
+no strict 'subs';
+eval q{ read($buffer, FILE, 1); };
+like $@, qr/^Can't modify constant item in read /,
+    'read rejects a bareword buffer as a constant item';
+
+done_testing;