+++ src/test/resources/unit/substitution_empty_braced_interpolation.t
@@ -1,0 +1,10 @@
+#!/usr/bin/env perl
+use strict;
+use warnings;
+use Test::More;
+
+eval q{{s//${}/; //}};
+like $@, qr/syntax error/,
+    'empty braced interpolation is invalid in a substitution replacement';
+
+done_testing;